----Update Business Logic to 2----
Implemented a class that stores information for a recipe. Also made a utility class that converts recipes to string with name on the first line, then the second line being a comma seperated list of ingredients. Then implemented a class that writes a recipe to file if the file does not already contain a recipe with the same name.